makes patron a namespace
