+++ lib/patron/student.rb
@@ -1,0 +1,4 @@
+class Patron
+  class Student < Patron
+  end
+end
