--- lib/student.rb
@@ -1,2 +1,0 @@
-class Student < Patron
-end
